bug fixing
Found a potential bug with landmark detection in monocular video tracking:
The parallel processing (default 8 threads) will split the frames into chunks and run landmark detector in parallel. However, when the landmarks were stored to the final array, these chunks are mingled causing misaligned landmarks with images.
My solution is straightforward: in config/base.py I set the num of parallel workers to 1, but for nersemble I set to 8 (as default).